refactor(progressView): move chainStreamWrite to dedicated writeChaining module
Addresses Cursor Bugbot review: chainStreamWrite is a write-queueing /
concurrency utility, not a serialization helper, so housing it in
serializationUtils.ts made the module doc comment misleading and the file
name a poor fit. Move it to a new writeChaining.ts module and update the
three importers.
